+++ node/clients/profile/profileV2.ts
@@ -168,6 +168,7 @@ export class ProfileClientV2 extends JanusClient {
       undefined,
       piiRequest
     )
+
     return this.get<Address[]>(url, {
       metric: 'profile-system-v2-getUserAddresses',
     })
@@ -186,6 +187,7 @@ export class ProfileClientV2 extends JanusClient {
   private translateToV1Address = (addresses: AddressV2[]) =>
     addresses.map((address: AddressV2) => {
       const addressV2 = address.document
+
       return {
         addressName: addressV2.addressName ?? addressV2.name,
         name: addressV2.name ?? addressV2.addressName ?? '',
@@ -235,6 +237,7 @@ export class ProfileClientV2 extends JanusClient {
     return Object.entries<string>(addressesObj).map(([key, stringifiedObj]) => {
       try {
         const address = JSON.parse(stringifiedObj) as Address
+
         address.id = key
 
         return address
@@ -252,16 +255,19 @@ export class ProfileClientV2 extends JanusClient {
         return addr
       }
     )
+
     const addressesV2 = this.translateToV2Address(addressesV1)
     const [toChange] = addressesV2.filter(
       (addr) => addr.id === addressesV1[0].id
     )
+
     return this.getProfileInfo(user).then((profile) => {
       return this.getUserAddresses(user, profile).then(
         (addresses: Address[]) => {
           const [address] = addresses.filter(
             (addr) => addr.addressName === addressesV1[0].id
           )
+
           if (address) {
             return this.patch(
               `${this.baseUrl}/${profile.id}/addresses/${address.id}`,
@@ -271,6 +277,7 @@ export class ProfileClientV2 extends JanusClient {
               }
             )
           }
+
           return this.post(
             `${this.baseUrl}/${profile.id}/addresses`,
             toChange.document,